fix(build): warning-free -Wpedantic build; rewrite socket basic tests
- Logging macros are now pure variadic (LOG_X(...)), so format-only calls
  are valid ISO C++ under -Wpedantic; no-op variants expand to ((void)0).
- Fix HAVE_HTTP_DEBUG LOG_TRACE which failed to compile with no arguments.
- Cast socket handles for %llx trace logging (was UB with int fds).
- config.h: include <cstddef>/<cstdio> it depends on.
- Export ScopedSocket from net::common alongside Socket/SocketAddr.
- socket_addr_test: fix multi-char literal ']:' that silently tested ':'.
- socket_basic_test: replace EXPECT_TRUE(true) placeholders that leaked
  fds with real assertions (TCP/UDP loopback round trips, move/RAII
  semantics, socket options); skip IPv6 cases on hosts without IPv6.

+++ include/SocketsHpp/http/server/http_server.h
@@ -1023,7 +1023,7 @@ namespace http
         protected:
             virtual void onSocketAcceptable(Socket socket) override
             {
-                LOG_TRACE("HttpServer: accepting socket fd=0x%llx", socket.m_sock);
+                LOG_TRACE("HttpServer: accepting socket fd=0x%llx", static_cast<unsigned long long>(socket.m_sock));
                 assert(std::find(m_listeningSockets.begin(), m_listeningSockets.end(), socket) !=
                     m_listeningSockets.end());
 
@@ -1046,7 +1046,7 @@ namespace http
 
             virtual void onSocketReadable(Socket socket) override
             {
-                LOG_TRACE("HttpServer: reading socket fd=0x%llx", socket.m_sock);
+                LOG_TRACE("HttpServer: reading socket fd=0x%llx", static_cast<unsigned long long>(socket.m_sock));
                 assert(std::find(m_listeningSockets.begin(), m_listeningSockets.end(), socket) ==
                     m_listeningSockets.end());
 
@@ -1073,7 +1073,7 @@ namespace http
 
             virtual void onSocketWritable(Socket socket) override
             {
-                LOG_TRACE("HttpServer: writing socket fd=0x%llx", socket.m_sock);
+                LOG_TRACE("HttpServer: writing socket fd=0x%llx", static_cast<unsigned long long>(socket.m_sock));
 
                 assert(std::find(m_listeningSockets.begin(), m_listeningSockets.end(), socket) ==
                     m_listeningSockets.end());
@@ -1094,7 +1094,7 @@ namespace http
 
             virtual void onSocketClosed(Socket socket) override
             {
-                LOG_TRACE("HttpServer: closing socket fd=0x%llx", socket.m_sock);
+                LOG_TRACE("HttpServer: closing socket fd=0x%llx", static_cast<unsigned long long>(socket.m_sock));
                 assert(std::find(m_listeningSockets.begin(), m_listeningSockets.end(), socket) ==
                     m_listeningSockets.end());
